ecommerce/requirements.txt
asgiref==3.5.2
attrs==22.1.0
dj-database-url==1.0.0
Django==4.1.3
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.1
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
whitenoise==6.2.0
